fix steam solid boiler not accepting items (#1680)

+++ src/main/java/com/gregtechceu/gtceu/common/machine/trait/customlogic/CannerLogic.java
@@ -72,7 +72,7 @@ public class CannerLogic implements GTRecipeType.ICustomRecipeLogic {
                     }
                     fluidStack1 = fluidStack1.copy();
                     fluidStack1.setAmount(
-                            fluidHandlerItem.fill(new FluidStack(fluidStack1.getFluid(), (int) fluidStack1.getAmount()),
+                            fluidHandlerItem.fill(new FluidStack(fluidStack1.getFluid(), fluidStack1.getAmount()),
                                     IFluidHandler.FluidAction.EXECUTE));
                     if (fluidStack1.getAmount() > 0) {
                         return GTRecipeTypes.CANNER_RECIPES.recipeBuilder(GTStringUtils.itemStackToString(item))
